Juggler release
Refactores main class into `Juggler` to resolve conflicts with php7's
`String` class. Adds composer and phpunit.
