+++ .gitignore
@@ -1,1 +1,5 @@
-# Created by .ignore support plugin (hsz.mobi)
+# Dependencies
+/vendor
+
+# Generated
+composer.lock
